[REFACTOR] 임대 주택 개수 타입 통일 #18

+++ src/main/java/com/ssafy/bango/domain/rentalhouse/dto/response/DongSummaryResponse.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 public class DongSummaryResponse {
     private String dongCode;
     private String dongName;
-    private Integer rentalCount;
+    private Long rentalCount;
     private Double avgLatitude;
     private Double avgLongitude;
 }
